+++ CHANGELOG.md
@@ -4,14 +4,12 @@
 - Ongoing polls a user has not yet voted in are displayed on the dashboard. !2349 @AntonBallmaier
 - Provide RestApi to edit store !2515 !2625 @KoeMai
 
-
 ## Changes
 - Make menubar sticky on mobile !2411 @AntonBallmaier
 - Added a new content page for it infos to menu !2590 @chriswalg
 - Removed freshdesk link to Support and renamed contact to contact & support !2590 @chriswalg
 - Changed persistent cookie timespan from 14 days to 38 years !2626 @chriswalg
 - Updated the number of registered users on the start page !2629 @alex.simm
-
 - Added home region and verified columns to member tables of regions and working groups !2602 #1548 @alex.simm
 
 ## Bugfixes
@@ -39,6 +37,7 @@
 - New vue component for address search !2632 @alex.simm
 - Replaced get store information to rest api !1449 @KoeMai
 - Replaced the history dialogs on profile pages with a vue component !2587 @alex.simm
+- Moved verification to memberlist !2646 @chriswalg
 
 ## Dev/Test/CI stuff
 - Fix given null parameter !2550 @BibaltiK